refactor: create search models submodule
